backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
pydantic==2.10.4
httpx==0.28.1
python-multipart==0.0.20
PyPDF2==3.0.1
python-dotenv==1.0.1
google-generativeai==0.8.4
